// frontend/lib/client/types.gen.ts
export type TripPublic = {
    /**
     * Title
     */
    title: string;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Location
     */
    location: string;
    /**
     * Date
     */
    date: string;
    /**
     * Cost
     */
    cost: string;
    /**
     * Max Students
     */
    max_students?: number;
    /**
     * Id
     */
    id: string;
    /**
     * School Id
     */
    school_id: string;
    /**
     * Registration Count
     */
    registration_count?: number | null;
    /**
     * Created At
     */
    created_at?: string | null;
};